Unify tests
